Refactor destination CP test coverage
